add program public templating

+++ fitness-app/server/programTemplate/deleteTemplate.ts
@@ -1,13 +1,24 @@
-"use server"
+"use server";
 
-import { prisma } from "@/lib/prisma"
+import { prisma } from "@/lib/prisma";
 
-export async function deleteTemplate(templateId : string) {
+export async function deleteTemplate(templateId: string) {
+  const template = await prisma.programTemplate.findUnique({
+    where: { id: templateId },
+    select: { sptemplate: true },
+  });
 
-    await prisma.programTemplate.delete({
-        where: {
-            id: templateId
-        }
-    })
+  if (!template) {
+    throw new Error("Template not found.");
+  }
 
+  if (template.sptemplate === true) {
+    throw new Error("SP templates cannot be deleted.");
+  }
+
+  await prisma.programTemplate.delete({
+    where: { id: templateId },
+  });
+
+  return { success: true };
 }
